+++ instrumentation/aws-lambda/aws-lambda-core-1.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/awslambdacore/v1_0/AwsLambdaRequestStreamHandlerInstrumentation.java
@@ -25,6 +25,7 @@ import io.opentelemetry.javaagent.extension.instrumentation.TypeTransformer;
 import java.io.InputStream;
 import java.util.Collections;
 import java.util.concurrent.TimeUnit;
+import javax.annotation.Nullable;
 import net.bytebuddy.asm.Advice;
 import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.matcher.ElementMatcher;
@@ -60,38 +61,58 @@ public class AwsLambdaRequestStreamHandlerInstrumentation implements TypeInstrum
   @SuppressWarnings("unused")
   public static class HandleRequestAdvice {
 
-    @Advice.OnMethodEnter(suppress = Throwable.class)
-    public static void onEnter(
-        @Advice.Argument(0) InputStream input,
-        @Advice.Argument(2) Context context,
-        @Advice.Local("otelInput") AwsLambdaRequest otelInput,
-        @Advice.Local("otelContext") io.opentelemetry.context.Context otelContext,
-        @Advice.Local("otelScope") Scope otelScope) {
+    public static class AdviceScope {
+      private final AwsLambdaRequest lambdaRequest;
+      private final io.opentelemetry.context.Context context;
+      private final Scope scope;
 
-      otelInput = AwsLambdaRequest.create(context, input, Collections.emptyMap());
-      io.opentelemetry.context.Context parentContext = functionInstrumenter().extract(otelInput);
-
-      if (!functionInstrumenter().shouldStart(parentContext, otelInput)) {
-        return;
+      private AdviceScope(
+          AwsLambdaRequest lambdaRequest,
+          io.opentelemetry.context.Context otelContext,
+          Scope scope) {
+        this.lambdaRequest = lambdaRequest;
+        this.context = otelContext;
+        this.scope = scope;
       }
 
-      otelContext = functionInstrumenter().start(parentContext, otelInput);
-      otelScope = otelContext.makeCurrent();
+      @Nullable
+      public static AdviceScope start(InputStream input, Context context) {
+        AwsLambdaRequest lambdaRequest =
+            AwsLambdaRequest.create(context, input, Collections.emptyMap());
+        io.opentelemetry.context.Context parentContext =
+            functionInstrumenter().extract(lambdaRequest);
+
+        if (!functionInstrumenter().shouldStart(parentContext, lambdaRequest)) {
+          return null;
+        }
+
+        io.opentelemetry.context.Context otelContext =
+            functionInstrumenter().start(parentContext, lambdaRequest);
+        return new AdviceScope(lambdaRequest, otelContext, otelContext.makeCurrent());
+      }
+
+      public void end(@Nullable Throwable throwable) {
+        scope.close();
+        functionInstrumenter().end(context, lambdaRequest, null, throwable);
+
+        OpenTelemetrySdkAccess.forceFlush(flushTimeout().toNanos(), TimeUnit.NANOSECONDS);
+      }
+    }
+
+    @Nullable
+    @Advice.OnMethodEnter(suppress = Throwable.class)
+    public static AdviceScope onEnter(
+        @Advice.Argument(0) InputStream input, @Advice.Argument(2) Context context) {
+      return AdviceScope.start(input, context);
     }
 
     @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class)
     public static void stopSpan(
-        @Advice.Thrown Throwable throwable,
-        @Advice.Local("otelInput") AwsLambdaRequest input,
-        @Advice.Local("otelContext") io.opentelemetry.context.Context functionContext,
-        @Advice.Local("otelScope") Scope functionScope) {
-
-      if (functionScope != null) {
-        functionScope.close();
-        functionInstrumenter().end(functionContext, input, null, throwable);
+        @Advice.Thrown @Nullable Throwable throwable,
+        @Advice.Enter @Nullable AdviceScope adviceScope) {
+      if (adviceScope != null) {
+        adviceScope.end(throwable);
       }
-
-      OpenTelemetrySdkAccess.forceFlush(flushTimeout().toNanos(), TimeUnit.NANOSECONDS);
     }
   }
 }