+++ javaagent-tooling/src/main/java/io/opentelemetry/javaagent/tooling/ignore/AdditionalLibraryIgnoredTypesConfigurer.java
@@ -250,7 +250,8 @@ public class AdditionalLibraryIgnoredTypesConfigurer implements IgnoredTypesConf
         .ignoreClass("com.google.common.")
         .allowClass("com.google.common.util.concurrent.")
         .allowClass("com.google.common.base.internal.Finalizer")
-        .allowClass("com.google.common.base.Java8Usage$$Lambda");
+        .allowClass("com.google.common.base.Java8Usage$$Lambda")
+        .allowClass("com.google.common.eventbus.Subscriber$");
 
     builder
         .ignoreClass("com.google.inject.")